Lab-5/examples/fastapi-scaffold/pyproject.toml
[project]
name = "capstone-api"
version = "0.1.0"
description = "AI-powered capstone prototype — CS-AI-2025 Spring 2026"
requires-python = ">=3.11"

dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "openai>=1.50",
  "google-generativeai>=0.8",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
]

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
  "httpx>=0.27",   # for testing
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
]
